requirements.txt
django>=5.0
psycopg2-binary
gunicorn
whitenoise
dj-database-url
django-unfold
django-import-export
Pillow
django-storages
boto3